refactor(vectortile): move mixin to a file and reference where used
This was previously moved to resolve a circular dependency, but this seems to
be a better solution.
